+++ db/migrate/20180617164753_create_feature_registration_feature_flag.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class CreateFeatureRegistrationFeatureFlag < ActiveRecord::Migration[5.1]
+  include FlipperMigration
+
+  def up
+    create_flag :feature_registration, enabled: true
+  end
+
+  def down
+    destroy_flag :feature_registration
+  end
+end